# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iris API",
    "description": "Self-hosted media aggregator + streamer — internal household API.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/me/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "operationId": "get_preferences",
        "responses": {
          "200": {
            "description": "The caller's recommendation preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "operationId": "update_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved preferences (normalised)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unknown language in the request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PreferencesResponse": {
        "type": "object",
        "required": [
          "languages",
          "genres",
          "include_anime",
          "onboarding_completed"
        ],
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "include_anime": {
            "type": "boolean"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "onboarding_completed": {
            "type": "boolean"
          }
        }
      },
...
      "UpdatePreferencesRequest": {
        "type": "object",
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "include_anime": {
            "type": "boolean"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "onboarding_completed": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}